First successful ELT + minor refactoring
